+++ src/main/routes/features/dashboard/claimantDashboardController.ts
@@ -2,9 +2,8 @@ import {NextFunction, Request, RequestHandler, Response, Router} from 'express';
 import config from 'config';
 import {DASHBOARD_CLAIMANT_URL} from '../../urls';
 import {CivilServiceClient} from 'client/civilServiceClient';
-import {getClaimantNotifications} from 'services/dashboard/getDashboardContent';
 import {AppRequest} from 'common/models/AppRequest';
-import {getDashboardForm} from 'services/features/caseProgression/dashboardService';
+import {getDashboardForm, getNotifications} from 'services/dashboard/dashboardService';
 import {Claim} from 'models/claim';
 import {generateRedisKey, getCaseDataFromStore} from 'modules/draft-store/draftStoreService';
 
@@ -25,7 +24,7 @@ claimantDashboardController.get(DASHBOARD_CLAIMANT_URL, (async (req: Request, re
     const caseData = await civilServiceClient.retrieveClaimDetails(claimId, <AppRequest>req);
     caseData.caseProgression[hearing]= {[paymentInformation]: redisClaim.caseProgression?.hearing?.paymentInformation};
     Object.assign(claim, caseData);
-    const dashboardNotifications = await getClaimantNotifications(claimId, claim, lang);
+    const dashboardNotifications = await getNotifications(claimId, claim, lang);
     const dashboardTaskList = await getDashboardForm(claim,claimId);
     res.render(claimantDashboardViewPath, {claim, claimId, dashboardTaskList, dashboardNotifications});
   } catch (error) {
